Mondo node-header relationships: review follow-ups
Follow-up fixes from the third automated review pass on #1326, left
uncommitted after #1327 merged:

- Harden RO relation-label lookup with try/except so an unresolvable
  term degrades to no label instead of blanking the relationships block;
  precise `seen: set[tuple[str, str]]` annotation; document the benign
  lock-free cache race.
- Extract the duplicated node-relationships template into a shared
  SectionNodeRelationships.vue rendered by both the disease and
  non-disease node-header branches.
