+++ frontend/src/pages/node/SectionOverview.vue
@@ -52,26 +52,7 @@
           :frequency-label="frequencyLabel"
           :node="node"
         />
-        <!--
-          Mondo disease<->X relationships. These enter the KG with a generic
-          biolink:related_to predicate but retain the real relation (an RO term)
-          in original_predicate; the backend resolves that term's label from the
-          KG. Grouped by relation label so the RO term's meaning is revealed.
-        -->
-        <AppDetail
-          v-for="group in relationshipGroups"
-          :key="`rel-${group.label}`"
-          :title="group.label"
-          :full="true"
-        >
-          <AppFlex align-h="left">
-            <AppNodeBadge
-              v-for="entity in group.entities"
-              :key="entity.id"
-              :node="entity"
-            />
-          </AppFlex>
-        </AppDetail>
+        <SectionNodeRelationships :node="node" />
         <AppDetail
           :blank="!otherMappings.length"
           title="Equivalent disease concepts in other termiologies"
@@ -200,21 +181,7 @@
           <span>{{ nodeVersion.version || "unknown" }}</span>
         </AppDetail>
 
-        <!-- Mondo related_to relationships (RO original_predicate), grouped by relation label -->
-        <AppDetail
-          v-for="group in relationshipGroups"
-          :key="`rel-${group.label}`"
-          :title="group.label"
-          :full="true"
-        >
-          <AppFlex align-h="left">
-            <AppNodeBadge
-              v-for="entity in group.entities"
-              :key="entity.id"
-              :node="entity"
-            />
-          </AppFlex>
-        </AppDetail>
+        <SectionNodeRelationships :node="node" />
 
         <AppDetail
           :blank="!otherMappings.length"
@@ -254,7 +221,7 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import omit from "lodash/omit";
-import type { Entity, Node } from "@/api/model";
+import type { Node } from "@/api/model";
 import AppDetail from "@/components/AppDetail.vue";
 import AppDetails from "@/components/AppDetails.vue";
 import AppNodeBadge from "@/components/AppNodeBadge.vue";
@@ -263,35 +230,13 @@ import AppTagList from "@/components/AppTagList.vue";
 import { useClinicalResources } from "@/composables/use-clinical-resources";
 import { useSourceVersions } from "@/composables/use-source-versions";
 import SectionClinicalReources from "./SectionClinicalReources.vue";
+import SectionNodeRelationships from "./SectionNodeRelationships.vue";
 
 type Props = { node: Node };
 
 const { node } = defineProps<Props>();
 const isDiseaseNode = computed(() => node.category === "biolink:Disease");
 
-/**
- * Group the node's Mondo disease<->gene relationships by their relation label
- * so each RO relation (e.g. "has material basis in germline mutation in")
- * becomes a titled block. Falls back to the relation CURIE if the KG had no
- * label.
- */
-const relationshipGroups = computed(() => {
-  const groups = new Map<string, { label: string; entities: Entity[] }>();
-  for (const rel of node.node_relationships ?? []) {
-    if (!rel.related_entity) continue;
-    // Prefer the KG-resolved label; fall back to the relation CURIE so distinct
-    // unlabeled relations stay distinguishable (same chain for key and title).
-    const key = rel.relation_label || rel.relation || "Related to";
-    const label = key.charAt(0).toUpperCase() + key.slice(1);
-    let group = groups.get(key);
-    if (!group) {
-      group = { label, entities: [] };
-      groups.set(key, group);
-    }
-    group.entities.push(rel.related_entity);
-  }
-  return [...groups.values()];
-});
 /** separate out mappings into categories */
 const clinicalSynopsis = computed(
   () =>
